requirements.txt
pandas
openpyxl
python-docx
networkx
matplotlib
numpy
